Логика удалёния рецензии из сервиса ReviewService перенесена в контроллер ReviewController. В ReviewService добавлен метод "getById" для получения ReviewEntity по идентификатору.

+++ src/main/java/com/shuvi/cinema/controller/ReviewController.java
@@ -2,7 +2,11 @@ package com.shuvi.cinema.controller;
 
 import com.shuvi.cinema.controller.dto.review.ReviewCreateRequest;
 import com.shuvi.cinema.controller.dto.review.ReviewResponse;
+import com.shuvi.cinema.entity.ReviewEntity;
+import com.shuvi.cinema.entity.UserEntity;
+import com.shuvi.cinema.exception.AccessDenied;
 import com.shuvi.cinema.service.api.ReviewService;
+import com.shuvi.cinema.service.api.UserService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -27,6 +31,8 @@ import static com.shuvi.cinema.common.ResourceConstant.REVIEW_API_PATH;
 @Tag(name = "Review", description = "Manager Review")
 public class ReviewController {
 
+    private final UserService userService;
+
     private final ReviewService reviewService;
 
     @PostMapping
@@ -43,6 +49,20 @@ public class ReviewController {
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteById(@NonNull @PathVariable UUID id) {
-        reviewService.deleteById(id);
+        final ReviewEntity reviewEntity = reviewService.getById(id);
+        final UserEntity user = userService.getCurrentUser();
+
+        if (reviewEntity.getAuthor().getId().equals(user.getId())) {
+            reviewService.deleteById(id);
+            return;
+        }
+
+        if (user.getRoles().stream().anyMatch((role) -> role.getName().equals("ROLE_ADMIN"))) {
+            reviewService.deleteById(id);
+            return;
+        }
+
+        throw new AccessDenied();
+
     }
 }
